Rename diversity_reencrypt command to diversity_data_reencrypt
