+++ cla_backend/apps/cla_butler/tests/test_command.py
@@ -2,7 +2,7 @@ import mock
 from django.test import TestCase
 from django.core import management
 from cla_butler.tests.mixins import CreateSampleDiversityData
-from cla_butler.management.commands.diversity_reencrypt import Command as ClaButlerDiversityReencryptCommand
+from cla_butler.management.commands.diversity_data_reencrypt import Command as ClaButlerDiversityReencryptCommand
 from cla_butler.models import DiversityDataCheck, ACTION, STATUS
 
 
@@ -25,9 +25,9 @@ class DiversityCheckCommandTestCase(CreateSampleDiversityData, TestCase):
 
 class DiversityReencryptCommandTestCase(CreateSampleDiversityData, TestCase):
     @mock.patch(
-        "cla_butler.management.commands.diversity_reencrypt.Command.get_old_key_passphrase", return_value="cla"
+        "cla_butler.management.commands.diversity_data_reencrypt.Command.get_old_key_passphrase", return_value="cla"
     )
-    @mock.patch("cla_butler.management.commands.diversity_reencrypt.Command.schedule_tasks")
+    @mock.patch("cla_butler.management.commands.diversity_data_reencrypt.Command.schedule_tasks")
     @mock.patch.dict("os.environ", {"PREVIOUS_DIVERSITY_PRIVATE_KEY": "I am not empty"})
     def test_create_tasks_all(self, mock_schedule_tasks, get_old_key_passphrase):
         ClaButlerDiversityReencryptCommand.chunk_size = 2
@@ -37,17 +37,17 @@ class DiversityReencryptCommandTestCase(CreateSampleDiversityData, TestCase):
             self.assertEqual(len(tasks), tasks_count)
 
         mock_schedule_tasks.side_effect = schedule_tasks
-        management.call_command("diversity_reencrypt")
+        management.call_command("diversity_data_reencrypt")
         self.assertTrue(mock_schedule_tasks.called)
 
     @mock.patch(
-        "cla_butler.management.commands.diversity_reencrypt.Command.get_old_key_passphrase", return_value="cla"
+        "cla_butler.management.commands.diversity_data_reencrypt.Command.get_old_key_passphrase", return_value="cla"
     )
-    @mock.patch("cla_butler.management.commands.diversity_reencrypt.Command.schedule_tasks")
+    @mock.patch("cla_butler.management.commands.diversity_data_reencrypt.Command.schedule_tasks")
     @mock.patch.dict("os.environ", {"PREVIOUS_DIVERSITY_PRIVATE_KEY": "I am not empty"})
     def test_create_tasks_some(self, mock_schedule_tasks, get_old_key_passphrase):
         ClaButlerDiversityReencryptCommand.chunk_size = 2
-        # Mark the first record as processed so that we can test that diversity_reencrypt process does not attempt to
+        # Mark the first record as processed so that we can test that diversity_data_reencrypt process does not attempt to
         # reencrypt this record
         DiversityDataCheck.objects.create(
             personal_details_id=self.pd_records_ids[0], action=ACTION.REENCRYPT, status=STATUS.OK
@@ -66,5 +66,5 @@ class DiversityReencryptCommandTestCase(CreateSampleDiversityData, TestCase):
                 self.assertNotIn(self.pd_records_ids[0], task)
 
         mock_schedule_tasks.side_effect = schedule_tasks
-        management.call_command("diversity_reencrypt")
+        management.call_command("diversity_data_reencrypt")
         self.assertTrue(mock_schedule_tasks.called)
